Moved all modules to src directory
