+++ go.mod
@@ -1,3 +1,3 @@
-module main.go
+module ops0
 
 go 1.19
